% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top10.R
\name{top10}
\alias{top10}
\title{Title: Welcome to our package
This function helps clean and analyze user input data.}
\usage{
top10()
}
\value{
hi this is the value that the function returns
}
\description{
Title: Welcome to our package
This function helps clean and analyze user input data.
}
\examples{
here is an example
}


